Add help/info messages to display if the user asks

+++ tools/db.sh
@@ -11,6 +11,12 @@
 # --- start mysqld
 startDatabase() {
 
+  if [ $1 = "-h" -o $1 = "--help" ] ; then
+    highlight "startDatabase"
+    echo "  Starts an instance of MySQL database daemon"
+    return
+  fi
+
   echo "Starting MySQL daemon"
   $mysqld_cmd &> /dev/null &
   mysqld_pid=$!
@@ -26,6 +32,12 @@ startDatabase() {
 # --- stop mysqld
 stopDatabase() {
 
+  if [ $1 = "-h" -o $1 = "--help" ] ; then
+    highlight "stopDatabase"
+    echo "  Stops a running instance of MySQL daemon"
+    return
+  fi
+
   echo "Stopping MySQL daemon"
   kill $mysqld_pid
   if [ $? -eq 0 ] ; then
@@ -39,6 +51,13 @@ stopDatabase() {
 # --- create the database set in the environment ---
 createDatabase() {
 
+  if [ $1 = "-h" -o $1 = "--help" ] ; then
+    highlight "createDatabase"
+    echo "  Will create a new database called $ctms_db_name and give necessary"
+    echo "  permissions to $ctms_db_user"
+    return
+  fi
+
   echo "Creating Database $ctms_db_name"
   $mysql_cmd -uroot -p$mysql_root_pass -e "create database $ctms_db_name";
   $mysql_cmd -uroot -p$mysql_root_pass -e "grant ALL on ${ctms_db_name}.* to '$ctms_db_user' identified by '$ctms_db_pass'"
@@ -50,6 +69,12 @@ createDatabase() {
 # --- drop a database
 dropDatabase() {
 
+  if [ $1 = "-h" -o $1 = "--help" ] ; then
+    highlight "dropDatabase"
+    echo "  Will drop the existing database $ctms_db_name"
+    return
+  fi
+
   echo "Dropping Database $ctms_db_name"
   $mysql_cmd -uroot -p$mysql_root_pass -e "drop database $ctms_db_name";
   # TODO actually check for success
@@ -60,6 +85,12 @@ dropDatabase() {
 # --- Load the database schema
 loadSchema() {
 
+  if [ $1 = "-h" -o $1 = "--help" ] ; then
+    highlight "loadSchema"
+    echo "  Load the schema $schema_file into the database"
+    return
+  fi
+
   echo "Loading schema"
   $mysql_cmd -u$ctms_db_user -p$ctms_db_pass $ctms_db_name < $schema_file
   success
@@ -68,6 +99,12 @@ loadSchema() {
 
 # --- Load the seeddata file
 loadSeedData() {
+
+  if [ $1 = "-h" -o $1 = "--help" ] ; then
+    highlight "loadSeedData"
+    echo "  Load necessary seed data from $seed_data_file into the database"
+    return
+  fi
 
   echo "Loading seed data"
   $mysql_cmd -u$ctms_db_user -p$ctms_db_pass $ctms_db_name < $seed_data_file
@@ -78,6 +115,13 @@ loadSeedData() {
 # --- dump the database to the file specified (or auto generate the name
 #     based on the current date and time)
 dumpDb() {
+
+  if [ $1 = "-h" -o $1 = "--help" ] ; then
+    highlight "dumpDb [file]"
+    echo "  Create a backup of the database $ctms_db_name and save it either in"
+    echo "  [file] or a file based on todays date"
+    return
+  fi
 
   bak="$1"
   if [ -z "$bak" ] ; then
